Split examples in multiple files and minor fixes to types
