+++ frontend/index.d.ts
@@ -9,8 +9,8 @@ export declare class Pool<T> {
 
   deposit(amount: BigInteger, batchSize: BigInteger): Promise<IncompleteDeposit | null>;
   completeDeposit(incompleteDeposit: IncompleteDeposit, batchSize: BigInteger): Promise<IncompleteDeposit | null>;
-  close(batchSize: BigInteger, idxArray: int[]): Promise<int[]>;
-  completeClose(incompleteClose: IncompleteClose, batchSize: BigInteger): Promise<IncompleteDeposit | null>;
+  close(batchSize: BigInteger): Promise<IncompleteClose | null>;
+  completeClose(incompleteClose: IncompleteClose, batchSize: BigInteger): Promise<IncompleteClose | null>;
   userStake(amount: BigInteger): Promise<any>;
   userWithdraw(): Promise<any>;
   getAssocList(): Promise<EntryList>;
