+++ frontend/ts-example/tsconfig.json
@@ -6,5 +6,5 @@
     "preserveConstEnums": true,
     "sourceMap": true
   },
-  "files": ["index.ts"]
+  "include": ["index.ts", "examples/*.ts"]
 }
